% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_imputationDS.R
\name{global_imputationDS}
\alias{global_imputationDS}
\title{Global imputation of missing data using MICE}
\usage{
global_imputationDS(df, id_col, m, maxit)
}
\arguments{
\item{df}{A data frame containing the variables to be imputed. Both numeric and
categorical variables are supported.}

\item{id_col}{A string specifying the name of the identifier column.
This column is excluded from the imputation process and used to aggregate the
results across multiple imputations.}

\item{m}{Number of imputations to generate (passed to \code{mice}).}

\item{maxit}{Maximum number of MICE iterations per imputation chain.}
}
\value{
A data frame where missing values have been imputed using the MICE
}
\description{
Perform a global imputation of missing values in a data frame using
the Multiple Imputation by Chained Equations (MICE) algorithm. The function
automatically identifies continuous and categorical variables, assigns appropriate
imputation methods (pmm, logreg, polyreg), and returns a single imputed data frame.
}
