% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_columnsDS.r
\name{remove_columnsDS}
\alias{remove_columnsDS}
\title{Remove specified columns from a data frame}
\usage{
remove_columnsDS(df, col_names)
}
\arguments{
\item{df}{A data frame from which columns will be removed.}

\item{col_names}{A character vector specifying the names of the columns to be removed.}
}
\value{
The input data frame with the specified columns removed.
}
\description{
Remove one or more columns from a data frame based on their names.
}
